Deploy new V3 contract with uint16 free tournament limit (0x0e2663b0)
Deployed ChessBotsTournamentV3 at 0x0e2663b0DCD9b7408d51C6972f679B81a5A7477e
on Monad mainnet. Updated contract address across all services, bots, web app,
and docs. Updated ABI types from uint8 to uint16 for sponsoredFreeTournaments
and maxFreeTournaments. Bumped auto-limit ceiling from 255 to 65535.

+++ services/tournament-orchestrator/src/chains/client.ts
@@ -133,15 +133,15 @@ const TOURNAMENT_ABI = [
       { name: 'totalTournaments', type: 'uint64' },
       { name: 'totalPrizeDistributed', type: 'uint64' },
       { name: 'paused', type: 'bool' },
-      { name: 'sponsoredFreeTournaments', type: 'uint8' },
-      { name: 'maxFreeTournaments', type: 'uint8' },
+      { name: 'sponsoredFreeTournaments', type: 'uint16' },
+      { name: 'maxFreeTournaments', type: 'uint16' },
     ],
     stateMutability: 'view',
     type: 'function',
   },
   // setFreeTournamentLimit (write) — authority only
   {
-    inputs: [{ name: '_newLimit', type: 'uint8' }],
+    inputs: [{ name: '_newLimit', type: 'uint16' }],
     name: 'setFreeTournamentLimit',
     outputs: [],
     stateMutability: 'nonpayable',
